Adaptive keyframe selection
Detects the level of difference between the current frame and the
previous frame, and if the difference is above a preset threshold,
encodes the current frame as a keyframe.

Introduces a new CLI option, --max-keyint (-i), to set the maximum
interval between two keyframes. Reuses the existing --keyint (-I) option
as the minimum interval between two keyframes. Sets the default minimum
to be 12 frames. To disable scenechange detection, set the max and min
to be equal via the CLI.