Make plane generic over pixel component type
In order to support both u8 and u16 for plane components, make the Plane
structure generic over the component type. As a consequence, many other
structures and functions also become generic.

Some functions are not u8-compatible yet, although they have been make
generic over the component type to make the compilation work. They
assert that the size of the generic parameter is 16 bits wide.

For this reason, the root context structure is unconditionally created
as Context<u16> for now.