Adapt Pixel trait for convert_slice_2d()
Since Pixel is only implemented for u8 and u16 types, it is always
possible to cast a Pixel to an integral primitive type. However, the
compiler could not know this, because in theory any other type could
also implement the Pixel trait.

To enforce this restriction, make the (main) integral primitive types
implement CastForPrimitive<T> for any T which implements Pixel. That
way, the compiler knows that a Pixel (whatever its concrete type) can
always be cast to an integer.

Then, rewrite the constraints on convert_slice_2d() to use
CastFromPrimitive, so that callers need not to add redundant constraints
if T implements Pixel.