Avoid use of Plane::p in Plane::downsample
Plane::p accesses pixels in a fashion that the compiler
cannot assume to be linear. This is harmful to both
compiler optimization and cache linearity.

Although we cannot completely eliminate this problem
since two rows must be accessed at a time for downsampling,
we can alleviate it by accessing the data slice directly.

Improves performance of av-scenechange by 15%.
Did not benchmark rav1e but presumably helps by 1-2% there also.